Move custom name logic to the base entity class
All entities have custom names.  Doing this removes a special case used
for players, and simplifies the process for the next change I'll make.

Note that players do NOT use the custom name tag to display their name;
the name tag you see in game is based off of their profile.  They still
have a special case for health, but that will be removed later.

This also affects the 1.8 protocol, which also lets all entities have
custom names in 1.8.  (I verified that the field was on entity in 1.8; it
was only on living entity on 1.7.10 though).