Create new entity metadata infrastructure
cMetadataWriter and its implementation cMetadataWriter190 are going to be
used to abstract entity metadata writing to the entities themselves.

Additionally, the protocol version enum was moved to Defines.h since
different protocol versions will need to be accessible to use
cMetadataWriter to its fullest, and protocols now have a member for this
version (again so that cMetadataWriter can use it).