Use cEntity::WriteMetadata for 1.9+ protocols
This does remove the giant enum in Protocol110x because we don't need it
anymore - this does the same thing as that enum (IE, automatic generation
of the metadata constants) without having to copy the entire metadata
writing code for each change.  It's also designed with the 1.11 changes in
mind, not just the 1.10 changes - it should be easy to expand for the
upcoming changes.