Add 1.11 entity metadata
This change adjusts the IDs to match the skeleton, zombie, and horse
changes made in 1.11.  However, it doesn't do so completely; it doesn't
handle the split entity IDs.  This means that zombie villagers, wither
skeletons, and horse variants all use their default versions (zombies,
skeletons, and regular horses).  But it should fix #3597.

Additionally, this corrects a mistake in 1.10's metadata writer, where
the ageable baby ID was used instead of the zombie baby ID.  Those
constants have the same value, so it wasn't a programatic issue, but it
was a semantic problem.