Changed cItemGrid::GetItem and similar functions to support size_t
As STL-Containers want to use unsigned indexes it is better to migrate
everything to size_t in the long run to avoid static_cast's.
Signed integers may still be used for special indexes, which may have some
special values, e.g. -1 if no index was found. This has the benefit that the
compiler will throw an error, if somebody always valid indexes and indexes that
may have a special value, as both distinguish in the signedness.