Temp fix for dissapearing section in 1.13 and 1.14 versions
Previously the value for nonEmptyBlocks in each chunk section was -1. This was probably set with the intention to set it to the max value. However, the game also interprets the value as signed. meaning when a block is placed the value is incremented to 0.  But when the nonEmptyBlocks value is 0 the game treats that section as empty and stops rendering blocks in it. 

This patch makes sure that the value sent is 4096, the amount of blocks in any section. This will probably lead to the inverse bug where the game will treat empty sections as non-empty.