Inital commit for new chunk format
Chunk generation is disabled is this commit
you have to manually copy and paste region files

Only blocks are supported. Biomes and heuight rmaps arent
Tile entites also are not supported

This system uses a ENUM_BLOCK to enemurate all block in the games in every blockstate possible.  Blocks are sorted alphabetically. Blockstates sorted alphabetically by state name.

This is neccessary because block ids are not consistent across versions.
In each version palette file there is supposed to be a function which translates the value from ENUM_BLOCK to the version's corresponding block id.

Many fucntions mainly related to block placing and breakig have been changed to work with the new system.

Loading / Saving in the new format
In this new foramt blocks are no longer stored directly as ids. The BlockStates long array is the actual block data. It is compromised of n-bit long values where each one is an index in to the palette. n is equal to Celi(Log2(palette_size)). The longs are stored in Big endian on disk.

In 1.16 and above if a value does not fit in a long it is stored in the next long insted of being split across two in older versions. According to the Wiki.

Every entry in the palette has a string named "Name" whose value is the block id. If the block has block states they are in a list "Properties". The ENUM_BLOCK value is acquired by passing the block id and its block states to then hash map "StringNameToId".  The string is in the following format. Block id with out "minecraft: " after it block states if any. after every block state  ", ". Block states are sorted alphabetically. After each block states name goes ": ".

Getting the string from an ENUM_BLOCKS value is done using the function PaletteUpgrade::GetSaveStrings the first element always has the block name. The rest if there are any are block states. The first element in each pair is the block state name and the second the value. In the first pair the first element is always an empty string and the second element in the pair is the block id.

When placing blocsk susually only the BlockType is known and the wanted block states. In order the get the ENUM_BLOCK value use the fuction PaletteUpgrade::ToInternalId which takes the BlockType and AllBlockStates struct
The AllBlockStates struct has every block states from every block that exist. Some have prefixes because of name conflicts. Changing block states which do not exist for the given block has not effect.

Compiler Problems
The reasons beacuse only dirt, grass block, bedrock and oak_trapdoor are implemented is because compile times balloon when the rest are added. I assume this is because MSVC struggles with compiling nearly 30k strings.
I have not tried other compilers.

Also if the hash map is initialized direclty in the header file like is it currently adding all the blocks this way requires extending the stack space. Otherwise an Stackoverflow exception is thown.