Add GitHub actions (#5611)
* Add GitHub actions

* Fix target branches for GH actions

* Change GH release name and tag

The name and tag of the uploaded release are now based on the branch name instead of the commit SHA

* Add experimental branch to GH actions

* Minor changes for GH actions

Co-authored-by: Mat <mail@mathias.is>

* Fixes for GH actions

* Binary uploading is now a separate job.
The ZIP files are first uploaded as artifacts and later in the other job redownloaded and uploaded to a release.
The tag for the release is updated when the new files are uploaded.
GH actions now won't run for pushes on tags that match the regex "release-*" this is to prevent GH actions being start in recursively.
Add 60 minute timeout to every job

* GH actions cleanup

* Add comment

* Disable release process

---------

Co-authored-by: Mat <mail@mathias.is>